Werror fixes
A bunch of fixes to get `-Werror` builds to pass,
both on Linux and on Windows (`/WX`)

* Disable verbose Windows warnings
  * `_CRT_SECURE_NO_WARNINGS` because of `std::getenv`
  * C4267 because of conversions from `size_t` to other integers
* Define `WIN32_LEAN_AND_MEAN` and `NOMINMAX` on Windows
  * Gets rid of some errors and speeds up the build
* Convert integer CUDA objects to std::uintptr_t
  before reinterpreting them to a pointer
* Fixed "unused function" warning for `GetHipFormatPixelSize`
  * There was a lot of duplication, now a single function
    called `imageElementByteSize`
* Mark some variables as potentially unused (only used in asserts)
* Other minor fixes

+++ test/adapters/cuda/urDeviceCreateWithNativeHandle.cpp
@@ -15,8 +15,8 @@ TEST_F(urCudaDeviceCreateWithNativeHandle, Success) {
     CUdevice cudaDevice;
     ASSERT_SUCCESS_CUDA(cuDeviceGet(&cudaDevice, 0));
 
-    ur_native_handle_t nativeCuda =
-        reinterpret_cast<ur_native_handle_t>(cudaDevice);
+    ur_native_handle_t nativeCuda = reinterpret_cast<ur_native_handle_t>(
+        static_cast<std::uintptr_t>(cudaDevice));
     ur_device_handle_t urDevice;
     ASSERT_SUCCESS(urDeviceCreateWithNativeHandle(nativeCuda, platform, nullptr,
                                                   &urDevice));
